made paypal live, but (check description)
but can't get tx id from random paypal user and this tx id is unique and essential to the database structure. gotta make a work around. it returned as null when the donation was made. but that's all that needs to be fixed. check models.py , perhaps we could use another value, just "id = db.Column(db.Integer, primary_key=True)" etc